+++ src/components/Layout/layout.js
@@ -1,4 +1,5 @@
 import s from "./style.module.css";
+import cx from "classnames";
 
 const Layout = ({ title, urlBg, colorBg, children }) => {
   const styleLayout = {};
@@ -17,7 +18,7 @@ const Layout = ({ title, urlBg, colorBg, children }) => {
             <h3>{title}</h3>
             <span className={s.separator}></span>
           </div>
-          <div className={`${s.desc} ${s.full}`}>
+          <div className={cx(s.desc, s.full)}>
             <p>{children}</p>
           </div>
         </article>
