+++ src/components/PokemonCard/pokemonCard.js
@@ -1,5 +1,6 @@
-import s from "./style.module.css";
 import { useState } from "react";
+import cx from "classnames";
+import s from "./style.module.css";
 
 import backImage from "./img/card-back-side.jpg";
 
@@ -10,15 +11,15 @@ const PokemonCard = ({ id, name, type, img, values }) => {
 
   return (
     <div className={s.root}>
-      <div className={`${s.pokemonCard} ${isActive ? s.active : null}`} onClick = {pokemonClickHandler}>
+      <div className={cx(s.pokemonCard, {[s.active]: isActive})} onClick = {pokemonClickHandler}>
         <div className={s.cardFront}>
-          <div className={`${s.wrap} ${s.front}`}>
-            <div className={`${s.pokemon} ${s.[type]}`}>
+          <div className={cx(s.wrap, s.front)}>
+            <div className={cx(s.pokemon, s.[type])}>
               <div className={s.values}>
-                <div className={`${s.count} ${s.top}`}>{values.top}</div>
-                <div className={`${s.count} ${s.right}`}>{values.right}</div>
-                <div className={`${s.count} ${s.bottom}`}>{values.bottom}</div>
-                <div className={`${s.count} ${s.left}`}>{values.left}</div>
+                <div className={cx(s.count, s.top)}>{values.top}</div>
+                <div className={cx(s.count, s.right)}>{values.right}</div>
+                <div className={cx(s.count, s.bottom)}>{values.bottom}</div>
+                <div className={cx(s.count, s.left)}>{values.left}</div>
               </div>
               <div className={s.imgContainer}>
                 <img src={img} alt={name} />
@@ -35,7 +36,7 @@ const PokemonCard = ({ id, name, type, img, values }) => {
         </div>
 
         <div className={s.cardBack}>
-          <div className={`${s.wrap} ${s.back}`}>
+          <div className={cx(s.wrap, s.back)}>
             <img src={backImage} alt="Сard Backed" />
           </div>
         </div>
